task: Add user authorization check to the products routes

+++ api/controllers/products.js
@@ -36,8 +36,11 @@ exports.products_create = (req, res, next) => {
         _id: new mongoose.Types.ObjectId,
         name: req.body.name,
         price: req.body.price,
-        productImage: req.file.path
+        user: req.userData.userId
     })
+    if (req.file !== undefined) {
+        product.productImage = req.file.path
+    }
     product.save()
         .then(({ _id, name, price }) => {
             return res.status(201).json({
@@ -82,35 +85,60 @@ exports.products_get_one = (req, res, next) => {
 }
 
 exports.products_edit = (req, res, next) => {
-    const id = req.params.productId
     const props = req.body
-    console.log(id)
-    Product.update({ _id: id }, props)
-        .select('-__v')
-        .exec()
-        .then(doc => {
-            return res.status(200).json({
-                msg: "Product updated",
-                request: {
-                    type: 'GET',
-                    url: `${PRODUCTS_URL}${id}`
-                }
+    const userId = req.userData.userId
+    const id = req.params.productId
+    // Check if the product with the same user exists
+    // Use 'find()' with 'limit()' rather than 'findOne' because 'findOne' returns the doc,
+    // slowing query. Find doesn't
+    if (Product.find({ _id: id, user: userId }).limit(1)) {
+        Product.update({ _id: id }, props)
+            .select('-__v')
+            .exec()
+            .then(doc => {
+                return res.status(200).json({
+                    msg: "Product updated",
+                    request: {
+                        type: 'GET',
+                        url: `${PRODUCTS_URL}${id}`
+                    }
+                })
+            }).catch(err => {
+                return res.status(500).json({ error: err })
             })
-        }).catch(err => {
-            return res.status(500).json({ error: err })
+    } else {
+        res.status(403).json({
+            message: "You are not authorized to edit this product"
         })
+    }
 }
 
 exports.products_delete = (req, res, next) => {
+    const userId = req.userData.userId
     const id = req.params.productId
-    Product.findByIdAndDelete(id)
-        .exec()
-        .then(result => {
-            return res.status(200).json({
-                msg: "Deletion successful",
-                url: PRODUCTS_URL
+    // Check if the product with the same user exists
+    // Use 'find()' with 'limit()' rather than 'findOne' because 'findOne' returns the doc,
+    // slowing query. Find doesn't
+    Product.exists({ _id: id, user: userId }, (err, result) => {
+        if (err) {
+            res.status(500).json({
+                error: err
             })
-        }).catch(err => {
-            return res.status(500).json({ error: err })
-        })
+        } else if (result) {
+            Product.findByIdAndDelete(id)
+                .exec()
+                .then(result => {
+                    return res.status(200).json({
+                        msg: "Deletion successful",
+                        url: PRODUCTS_URL
+                    })
+                }).catch(err => {
+                    return res.status(500).json({ error: err })
+                })
+        } else {
+            res.status(401).json({
+                message: "You do not have permission to delete that item"
+            })
+        }
+    })
 }
